chore: create new project for streaming example
