+++ src/IBM.Watson.TextToSpeech.v1/Examples/ServiceExample.cs
@@ -341,46 +341,5 @@ namespace IBM.Watson.TextToSpeech.v1.Examples
             Console.WriteLine(result.StatusCode);
         }
         #endregion
-
-        #region Websockets
-        /*
-        public void SynthesizeUsingWebsockets()
-        {
-            IamAuthenticator authenticator = new IamAuthenticator("{apikey}");
-            TextToSpeechService service = new TextToSpeechService(authenticator);
-
-            SynthesizeCallback callback = new SynthesizeCallback();
-
-            MemoryStream soundStream = new MemoryStream();
-            // Example requires SoundPlayer
-            SoundPlayer player = new SoundPlayer(soundStream);
-
-            callback.OnOpen = () =>
-            {
-                Console.WriteLine("open");
-            };
-
-            callback.OnClose = () =>
-            {
-                Console.WriteLine("close");
-            };
-            callback.OnMessage = (bytes) =>
-            {
-                player.Stream.Position = player.Stream.Length;
-                player.Stream.WriteAsync(bytes, 0, bytes.Length);
-                WaveUtils.ReWriteWaveHeader((MemoryStream)player.Stream);
-                Console.WriteLine("new message call");
-            };
-
-            var synthesizeResult = service.SynthesizeUsingWebsockets(
-                voice: SynthesizeEnums.VoiceValue.EN_US_ALLISONVOICE,
-                callback: callback,
-                accept: SynthesizeEnums.AcceptValue.AUDIO_WAV,
-                timings: new string[] { "words" },
-                text: "Lorem Ipsum is simply dummy text of the printing and typesetting industry.");
-            player.PlaySync();
-        }
-        */
-        #endregion
     }
 }
