+++ src/IBM.Watson.TextToSpeech.v1/Examples/Streaming/IBM.Watson.TextToSpeech.v1.Examples.Streaming.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+  
+  <ItemGroup>
+    <PackageReference Include="IBM.Cloud.SDK.Core" Version="1.2.0" />
+  </ItemGroup>
+  
+  <ItemGroup>
+    <ProjectReference Include="..\..\IBM.Watson.TextToSpeech.v1.csproj" />
+  </ItemGroup>
+
+</Project>
